Converted skillcape customise widget to use typescript
