+++ src/main/javascript/modules/widget/index.ts
@@ -1,3 +1,4 @@
 import './makex';
 import './report';
 import './settings';
+import './skillcape-customiser';
